Unify the minimum height of most element  (#8797)
## Describe your changes

This PR unifies most of our elements to a minimum height of 2.5rem
(40px) and also contains a couple of other papercuts and theming-related
refactorings, e.g.:

- a lot of widgets get downsized from 40.4px to 40px
- Color picker & single-line radio get upsized to 40px
- add more roundness to checkbox
- remove left margin for toggle, checkmark, and radio
- remove padding of table
- Use variables for some shared values (e.g., border width). The current
state of theming feels quite messy, and eventually, it should get a
proper refactoring. But I think it's a good first step to move some of
those style props to shared variables to simplify a later refactoring.
- Update playwright to 1.44 (since we have to update a large chunk of
screenshot. It's a good moment to also update this, which also requires
updating lots of screenshots)

## GitHub Issue Link (if applicable)

- Closes https://github.com/streamlit/streamlit/issues/8027
- Closes https://github.com/streamlit/streamlit/issues/8706

## Testing Plan

- No logic changed, updated e2e snapshots.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.