Improvements to NumberInput formatting (#9035)
## Describe your changes

This PR changes a few things related to the formatting of the
`<NumberInput />` component:
1. The docs were updated to provide more information about the `format`
parameter and how to use it.
2. If the number is a float, a `step` is specified (and is a float), and
a `format` is not specified, we format the result with the same number
of decimal places as the `step`.
3. Unit tests to cover newly added functionality.

## GitHub Issue Link (if applicable)
https://github.com/streamlit/streamlit/issues/7163

## Testing Plan

These changes have been tested via jest unit tests. Specifically, it
adds tests for:
- normal case of using a formatting string to set decimal places
- using a formatting string to format a float as an integer
- automatically formatting when the condition above is met (float +
float `step` + no `format` provided)
- not automatically formatting when explicit `format` is provided
- not automatically formatting when the `step` is an integer

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.