Utilize null or undefined helpers & add lint rule (#9153)
## Describe your changes

This PR is a followup of [this
comment](https://github.com/streamlit/streamlit/pull/9096#discussion_r1682563094)
where we will now use the type guarded helpers `isNullOrUndefined` and
`notNullOrUndefined` instead of `== null/undefined` or `!=
null/undefined`.

## Testing Plan
If tests pass, I assume this worked as expected. 


---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.