Enable a little additional strictness in mypy.ini (#9181)
In mypy.ini, all of the features of strict mode are listed, apparently
being gradually uncommented as the codebase comes into compliance with
them. Presumably eventually the body of them will be replaced with
strict=true. But this is not that day. This change just enables two of
them, because the code already passes those and they don't have to be
commented. It also stops ignoring errors in add_license_headers, because
that file doesn't even exist anymore.