Update websocket reconnect e2e test to use the data-test-connection-state attribute (#9426)
## Describe your changes

We have observed some flakiness with our websocket e2e tests. This PR
updates the observer code to leverage the newly added `connection-state`
attribute (https://github.com/streamlit/streamlit/pull/9131); since the
observer is "simpler" in the sense of that it observes the attributes of
a static node instead of iterating through a potentially changing node
list, we hope to also fix the flakiness. The assumption is that
sometimes the DOM changes were not observed because they were collapsed
by React or so.

## GitHub Issue Link (if applicable)

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- update the existing test to leverage the new connection-state
attribute
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.