Replace ad-hoc computation of reversed session_state map with in-memory version (#9472)
## Describe your changes

Closes #9415 

In #9415, we have discovered that the current implementation of
Streamlit's `session_state.filtered_state` property / function is kind
of slow, because the reverse-map is computed on-the-fly every time the
property is accessed. This PR introduces a wrapper-dict that keeps the
original dict as well as the reversed version in sync.

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
  - Add unit tests for the new wrapper dict
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.