Refactor vega chart tooltip CSS (#9587)
## Describe your changes

We currently have overlapping global CSS styles for vega tooltips in
`globalStyles` and in vega-tooltip.css. This change moves the styles
into the ArrowVegaLiteChart component instead and applies it as global
styles.

## Testing Plan

- Add e2e test to check styling.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.