Improve tests for AppNodes and Quiver (#9653)
## Describe your changes

In the move of the Vite conversion, these tests failed because vitest
has slightly more strict equality check Jest seems to worry about the
contents (ie the elements) whereas Vitest will test all the keys of the
object (since vectors aren't strict arrays). This change extracts
vectors to be an array at the end (or nested arrays) and allow the
equality measure to translate.

This change also mocks out some of the window location to be explicit.

## Testing Plan

- JS unit tests updated.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.