Cleanup global styles - part 3 (#9772)
## Describe your changes

In context of Advanced Theming, we need to get rid of most global styles
in order to be able to style the different parts of the apps (main,
sidebar, ...) depending on the theming information.
In this PR, we remove the headings directive from the global css (`h1`,
`h2`, ...)

## GitHub Issue Link (if applicable)

## Testing Plan

Only moving style directives. Existing snapshot tests should catch
potential differences, so no tests are needed.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.