Remove unused global styles (#9804)
## Describe your changes

This PR removes a couple of unused global styles. These where added and
used a long time ago, but the latest state is using styles that are
defined directly with the relevant components. This cleanup is related
to the advanced theming project.

## Testing Plan

- No logical or visual changes. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.