Cleanup even more global styles (#9861)
## Describe your changes

This PR removes a couple more unused global styles. These were added and
used a long time ago, but the latest state uses styles defined directly
with the relevant components. This cleanup is related to the advanced
theming project.

## Testing Plan

- No logical or visual changes. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.