Set Python 3.9 as minimum version and remove unused type-ignore comment (#9919)
## Describe your changes

Fixes an issue in our build pipeline when mypy is run (see
[here](https://github.com/streamlit/streamlit/actions/runs/12004642925))
by removing the comment. We set Python 3.9 as the new min as 3.8 would
still require the comment. We are about to bump the max version to 3.13
and want to retire 3.8: https://github.com/streamlit/streamlit/pull/9635

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.