Refactor dataframe sizing to use values responsive to root font size (#9887)
## Describe your changes

Extends the grid theme object to include all the relevant sizing
configurations, instead of having these size variables stored as
constants. This also calculates the pixel sizes based on rem to make the
table responsive to the root font size.

## Testing Plan

- Update tests. No logical or visual changes, existing snapshots should
be sufficient.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.