Update outdated snapshots to get CI working again (#10108)
## Describe your changes

https://github.com/streamlit/streamlit/pull/10047 merged in visual
changes to chat input. For unknown reasons, this caused three outdated
snapshots to exist in our repo. This should have been caught by the CI
checks in the PR, not sure why it did not work in this case...

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.