Clean up vendored ipython logic (#10134)
## Describe your changes

This PR moves `modified_sys_path` to the `exec_code` module. The
`modified_sys_path` is a small class inspired by a similar class in
Ipython. But for me, it does not seem like a vendored ependency that we
have to treat in any special way.

Also, our notices file only contains licenses from frontend packages.
It's a bit inconsistent that we include this one backend license in
notices as well.

## Testing Plan

- No logical changes -> existing tests should cover well. 

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.