Don't update FinishedMessage status of full app runs by fragment runs (#10132)
## Describe your changes

We stopped raising a `RuntimeError` when a fragment with a specified id cannot be found. This solves the original issue in https://github.com/streamlit/streamlit/issues/9921, but surfaced a different issue where the dialog did not close anymore when the button in the dialog was clicked in fast succession. The reason for this is that in `forward_msg_queue`, the `FinishedScript` type prior to this PR here is always set to `STOPPED_EARLY_FOR_RERUN`. This is now modified so that the status is not changed for full app reruns triggered by a fragment runs. This resolves the described issue with the not-closing dialog because a second, fast subsequent click on the button does not override the initial finished-message in the queue belonging to the `st.rerun()` in the example.
This is correct since a fragment indeed does not stop full app reruns. This only happens when the fragment run is so fast that the browser queue is not cleared yet.

## GitHub Issue Link (if applicable)

## Testing Plan

- Unit Tests (JS and/or Python)
  - Updated Python unit tests.
- E2E Tests
  - Added an e2e test that can catch an underlying issue where the dialog does not close correctly. However, since the cause of the particular issue addressed in the PR is a race condition, the test can succeed even if the issue exists. However, it should still fail from time to time which can help us to get aware of an underlying problem.
- Any manual testing needed
  - The test introduced in the e2e test is more likely to fail when executing manually, so that should be done if an issue arises.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this project are made under the Apache 2.0 license.