Don't request_rerun for fragments that do not exist anymore (#10147)
## Describe your changes

We stopped raising
a RuntimeError when a fragment with a specified id cannot be found. This
solves the original issue in
https://github.com/streamlit/streamlit/issues/9921, but surfaced a
different issue where the dialog did not always close when the button in
the dialog was clicked in fast succession.

The reason for this is that in app_session, a fragment run might create
a new ScriptRunner when the current ScriptRunner is in state STOPPED
is false and the new ScriptRunner is created). This will lead to all
events from the previous script runner being ignored.
When the full app rerun ScriptRunner is done (STOPPED) but its events
are not processed before the new ScriptRunner is created, its finished
message is not sent to the frontend and no cleanup is happening.

## GitHub Issue Link (if applicable)

## Testing Plan

- Unit Tests (JS and/or Python)
- Add unit test to ensure that the scriptrunner is not requested if the
fragment does not even exist anymore
- E2E Tests
- The e2e test that was added in
https://github.com/streamlit/streamlit/pull/10132 _could_ catch this one
here, although it happens more rarely than the race condition tackled in
that PR

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.