[feat] Utilize React 18 createRoot API (#10399)
## Describe your changes

Note: This is a PR into the long-running feature branch
`feature/react-create-root`, not `develop`! Failing tests are expected
while we work through the long tail of issues.

- This implements Phase 2 of the React 18 `createRoot` plan by
introducing just the minimally required bootstrapping logic to utilize
the new API.

## GitHub Issue Link (if applicable)

## Testing Plan

- Updates the test utils to ensure they are running in React 18 mode

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.