[fix] Tooltip positioning logic with createRoot (#10436)
## Describe your changes

Note: This is a PR into the long-running feature branch
`feature/react-create-root`, not `develop`! Failing tests are expected
while we work through the long tail of issues.

- Fixes an issue where the Tooltip would appear visually incorrect when
rendering in `createRoot`
- Due to the new batching rendering logic, there was a timing issue when
attempting to measure position of elements. To fix this, we keep some
local state for the Tooltip's state and defer its measurement into a
`useEffect`

## GitHub Issue Link (if applicable)

## Testing Plan

- ✅ This fixes e2e tests that were previously failing for Tooltip

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.