[fix] Dataframe test fixes with createRoot (#10483)
## Describe your changes

Note: This is a PR into the long-running feature branch
feature/react-create-root, not develop!

- Stabilizes as many dataframe-related tests as possible
- The biggest difference with dataframe and other tests is that Glide
uses a `canvas` under the hood. Given the changes with `createRoot`, the
timing of application of these changes caused many of the dataframe
tests (mostly those that deal with interaction) to be flaky.
- This PR includes a number of new test utils that stabilize the tests
by doing more intelligent waiting.
- As a last resort, some tests had to be skipped. Reasons for skipping
are written in comments in the code.
- This was the last big chunk of failing tests with `createRoot`, hence
why we actually have a clean build now!

## GitHub Issue Link (if applicable)

## Testing Plan

- ✅ This updates and stabilizes many e2e tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.